feat: set up nginx server for ssl and https (#1, #PR7)
use nginx template configuration for serving application and passing
DOMAIN value dynamically to the configuration.

refine nginx.conf default configuration to include what is necessary
for the application but leave everything else in the template.

update the docker compose template to bind the nginx.conf file and
the template folder to the nginx container

update the deploy workflow to set the DOMAIN value in the .env file

update the test workflow to run test against the development server