refactor: remove docker init redundancies and set app port to 8000
Application port is exposed internally only or through the service
domain in the docker network, so can easily set it to 8000 and not
worry about interference.

The nginx app will now redirect requests to the gunicorn app
server's 8000 port.

The gunicorn app is no longer exposed on the local machine when
running services via docker compose either. Only the nginx server
is exposed.